requirements.txt
telegraph
requests
pillow
shutilwhich
imageio
urllib3==1.25.11
